pyproject.toml
[project]
name = "task-notifier"
version = "0.1.0"
description = ""
authors = [
    {name = "Gautham P Raj",email = "gauthampraj2003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-socketio[asyncio-server] (>=5.13.0,<6.0.0)",
    "python-socketio[asyncio-server] (>=5.13.0,<6.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "task_notifier", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
